+++ .changelog/2469.internal.2.md
@@ -1,0 +1,1 @@
+client: Rename the rpc crate to enclave_rpc
