+++ client/src/enclave_rpc/api/enclaverpc.rs
