+++ client/src/enclave_rpc/client.rs
