+++ client/src/enclave_rpc/macros.rs
